pyproject.toml
[tool.poetry]
name = "stock-market-dwh"
version = "0.1.0"
description = ""
authors = ["plenzke <3njoyth3sil3nc399@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
